feat: draft 尝试重构 legend 组件
