+++ package.json
@@ -31,7 +31,7 @@
     "lint-fix:tslint": "tslint -c tslint.json --fix 'src/**/*.ts' 'tests/**/*'",
     "coverage": "jest --coverage",
     "test": "jest",
-    "test-live": "DEBUG_MODE=1 jest --watch tests",
+    "test-live": "DEBUG_MODE=1 jest --watch",
     "ci": "run-s build test coverage",
     "prepublishOnly": "npm-run-all --parallel test build"
   },
